+++ tests/test_dflowfm.py
@@ -2,6 +2,7 @@ from os.path import abspath, dirname, join
 from hydromt_delft3dfm import DFlowFMModel
 import numpy as np
 from pathlib import Path
+import pytest
 
 TESTDATADIR = join(dirname(abspath(__file__)), "data")
 EXAMPLEDIR = join(dirname(abspath(__file__)), "..", "examples")
@@ -21,37 +22,55 @@ def test_write_empty_model(tmpdir):
     mod1.write()
 
 
-def test_write_readonlymode(tmpdir, caplog):
+def test_init_dflowfmmodel_mode_write_crs_none(tmpdir):
+    """
+    tests whether the crs is parsed properly
+    https://github.com/Deltares/hydromt_delft3dfm/issues/247
+    """
     root = join(tmpdir, "dflowfm_example")
-    mod1 = DFlowFMModel(
-        root=root,
-        mode="w",
-        crs=3857,
-    )
-    mod1.write()
-    
-    mod2 = DFlowFMModel(
-        root=root,
-        mode="r",
-        crs=3857,
-    )
-    mod2.write()
-    assert "Cannot write in read-only mode" in caplog.text
+    with pytest.raises(ValueError) as e:
+        _ = DFlowFMModel(root=root, mode="w")
+    assert "crs argument cannot be None with mode" in str(e.value)
+    with pytest.raises(ValueError) as e:
+        _ = DFlowFMModel(root=root, mode="w+")
+    assert "crs argument cannot be None with mode" in str(e.value)
+
+
+def test_init_dflowfmmodel_mode_read_crs_none():
+    """
+    tests whether the crs is parsed properly
+    https://github.com/Deltares/hydromt_delft3dfm/issues/247
+    """
+    root = join(EXAMPLEDIR, "dflowfm_local")
+    model1 = DFlowFMModel(root=root, mode="r")
+    model2 = DFlowFMModel(root=root, mode="r+")
+    assert model1.crs.to_epsg() == 32647
+    assert model2.crs.to_epsg() == 32647
+
+
+def test_init_dflowfmmodel_mode_read_crs_notnone(tmpdir):
+    """
+    tests whether the crs is parsed properly
+    https://github.com/Deltares/hydromt_delft3dfm/issues/247
+    """
+    # TODO: the model crs is actually 32647, but is overwritten here
+    # https://github.com/Deltares/hydromt_delft3dfm/issues/119
+    root = join(EXAMPLEDIR, "dflowfm_local")
+    model3 = DFlowFMModel(root=root, mode="r", crs=4326)
+    model4 = DFlowFMModel(root=root, mode="r", crs=4326)
+    assert model3.crs.to_epsg() == 4326
+    assert model4.crs.to_epsg() == 4326
 
 
 def test_read_write_config_empty_paths(tmpdir):
     # Instantiate an empty model
     root = join(tmpdir, "dflowfm_example")
-    model1 = DFlowFMModel(
-        root=root,
-        mode="w",
-        crs=3857,
-    )
+    model1 = DFlowFMModel(root=root, mode="w", crs=3857)
     # Get the mdu settings
     model1.mdu.read()
     # Check whether the path is an emtpy string
     assert model1.mdu.data["output"]["outputdir"] == ""
-    
+
     # write the model to read it again
     model1.write()
     model2 = DFlowFMModel(root=root, mode="r", crs=3857)